requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
python-dotenv
google-generativeai
python-multipart
PyPDF2
python-docx